Refactor code for loadImage
